+++ java/src/Ice/src/main/java/com/zeroc/IceInternal/ConnectionFlushBatch.java
@@ -6,7 +6,7 @@ package com.zeroc.IceInternal;
 
 import java.util.concurrent.Callable;
 
-public class ConnectionFlushBatch extends OutgoingAsyncBaseI<Void> {
+public class ConnectionFlushBatch extends OutgoingAsyncBase<Void> {
   public ConnectionFlushBatch(
       com.zeroc.Ice.ConnectionI con, com.zeroc.Ice.Communicator communicator, Instance instance) {
     super(communicator, instance, "flushBatchRequests");
