+++ java/src/Ice/src/main/java/com/zeroc/IceInternal/OutgoingAsyncBase.java
@@ -5,30 +5,104 @@
 package com.zeroc.IceInternal;
 
 //
-// This non-generic base interface allows us to pass instances without causing "rawtypes" lint
-// warnings.
+// Base class for handling asynchronous invocations. This class is
+// responsible for the handling of the output stream and the child
+// invocation observer.
 //
-public interface OutgoingAsyncBase {
-  boolean sent();
+public abstract class OutgoingAsyncBase<T> extends InvocationFutureI<T> {
+  public boolean sent() {
+    return sent(true);
+  }
 
-  boolean completed(com.zeroc.Ice.InputStream is);
+  public boolean completed(com.zeroc.Ice.InputStream is) {
+    assert (false); // Must be implemented by classes that handle responses
+    return false;
+  }
 
-  boolean completed(com.zeroc.Ice.Exception ex);
+  public boolean completed(com.zeroc.Ice.Exception ex) {
+    return finished(ex);
+  }
 
-  void attachRemoteObserver(
-      com.zeroc.Ice.ConnectionInfo info, com.zeroc.Ice.Endpoint endpt, int requestId);
+  public final void attachRemoteObserver(
+      com.zeroc.Ice.ConnectionInfo info, com.zeroc.Ice.Endpoint endpt, int requestId) {
+    com.zeroc.Ice.Instrumentation.InvocationObserver observer = getObserver();
+    if (observer != null) {
+      final int size = _os.size() - Protocol.headerSize - 4;
+      _childObserver = observer.getRemoteObserver(info, endpt, requestId, size);
+      if (_childObserver != null) {
+        _childObserver.attach();
+      }
+    }
+  }
 
-  void attachCollocatedObserver(com.zeroc.Ice.ObjectAdapter adapter, int requestId);
+  public final void attachCollocatedObserver(com.zeroc.Ice.ObjectAdapter adapter, int requestId) {
+    com.zeroc.Ice.Instrumentation.InvocationObserver observer = getObserver();
+    if (observer != null) {
+      final int size = _os.size() - Protocol.headerSize - 4;
+      _childObserver = observer.getCollocatedObserver(adapter, requestId, size);
+      if (_childObserver != null) {
+        _childObserver.attach();
+      }
+    }
+  }
 
-  com.zeroc.Ice.OutputStream getOs();
+  public final com.zeroc.Ice.OutputStream getOs() {
+    return _os;
+  }
 
-  void invokeSent();
+  public T waitForResponse() {
+    if (Thread.interrupted()) {
+      throw new com.zeroc.Ice.OperationInterruptedException();
+    }
 
-  void invokeSentAsync();
+    try {
+      return get();
+    } catch (InterruptedException ex) {
+      throw new com.zeroc.Ice.OperationInterruptedException(ex);
+    } catch (java.util.concurrent.ExecutionException ee) {
+      try {
+        throw ee.getCause().fillInStackTrace();
+      } catch (RuntimeException ex) // Includes LocalException
+      {
+        throw ex;
+      } catch (Throwable ex) {
+        throw new com.zeroc.Ice.UnknownException(ex);
+      }
+    }
+  }
 
-  void invokeCompleted();
+  protected OutgoingAsyncBase(com.zeroc.Ice.Communicator com, Instance instance, String op) {
+    super(com, instance, op);
+    _os = new com.zeroc.Ice.OutputStream(instance, Protocol.currentProtocolEncoding);
+  }
 
-  void invokeCompletedAsync();
+  protected OutgoingAsyncBase(
+      com.zeroc.Ice.Communicator com, Instance instance, String op, com.zeroc.Ice.OutputStream os) {
+    super(com, instance, op);
+    _os = os;
+  }
 
-  void cancelable(CancellationHandler handler);
+  @Override
+  protected boolean sent(boolean done) {
+    if (done) {
+      if (_childObserver != null) {
+        _childObserver.detach();
+        _childObserver = null;
+      }
+    }
+    return super.sent(done);
+  }
+
+  @Override
+  protected boolean finished(com.zeroc.Ice.Exception ex) {
+    if (_childObserver != null) {
+      _childObserver.failed(ex.ice_id());
+      _childObserver.detach();
+      _childObserver = null;
+    }
+    return super.finished(ex);
+  }
+
+  protected com.zeroc.Ice.OutputStream _os;
+  protected com.zeroc.Ice.Instrumentation.ChildInvocationObserver _childObserver;
 }
